refactor: move workflow manager into dedicated module
